issue #17 COMBINED IF STATEMENT
I tried to think and think then I came to the conclusion. I can put all the four if statements in one if statement. My thinking is, either of the condition if its met then we shall have the true Boolean and it will not affect which condition to be excused.