issue #25 METHOD NAME & VARIABLE MODIFIED
The method name had a variable with a present or continuous tense. I came to realize names of methods or variables should be carefully considered since they carry proper information which can be used by another developer. Therefore the word Displaying was showing something in a present continous. Now t is just Display.

Also regarding the variable name or paramater John I can explain that I got it from mentor Rakete one day in a video call he was explaining to me a concept that any varible name can stay there and doesnt affect the argument which will be passed on. I removed it and put the reflecting one. That John variable name was for teaching purposes when Rakete was controlling my screen. Sorry I did not take action to change it.